# api-docs.yaml
openapi: 3.0.1
info:
  title: Clinfec
  description: test.
  contact:
    email: bieltdb20@gmail.com
  version: 1.0.0
...
servers:
  - url: https://localhost:3000/v1
  - url: http://localhost:3000/v1
...
paths:
  /users:
    post:
      tags:
        - User
      summary: Create user
      description: Creates a new User in the DataBase
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      security: []
      responses:
        200:
          description: User correctly added to database
        400:
          description: Error validating input data
        409:
          description: Username already exists, please pick another
        default:
          $ref: "#/components/responses/Default"

    get:
      tags:
        - User
      summary: Gets list of users(Only Admin) or User account details (Not Admin)
      description: Gets a list of all the existing users in the DataBase if the user
        sending the request is an admin, otherwise the response only targets the active
        user.
      operationId: getUserInfo
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Default"

    put:
      tags:
        - User
      summary: Edits active user data
      operationId: editActiveUserData
      requestBody:
        description: Fields to change
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/User"
      responses:
        200:
          description: Successful operation
        400:
          description: Invalid username/password supplied
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        409:
          description: Username or email already registered
        default:
          $ref: "#/components/responses/Default"

    delete:
      tags:
        - User
      summary: Disables user account (validation via token)
      operationId: deleteActiveUser
      responses:
        200:
          description: User account disabled
        401:
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Default"
  /users/login:
    get:
      tags:
        - User
      security: []
      summary: Logs user into the system
      operationId: loginUser
      parameters:
        - name: username or email
          in: query
          description: The user name or email for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        400:
          description: Invalid username/password supplied
        401:
          $ref: "#/components/responses/Disabled"
        default:
          $ref: "#/components/responses/Default"
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 100
        username:
          type: string
          example: ghee_dough
        password:
          type: string
          example: MySecretPass123
        email:
          type: string
          example: gheedough@gmail.com
        delivery_address:
          type: string
          example: Fake St. 123
        full_name:
          type: string
          example: Guido Torres
        phone:
          type: integer
          example: 5491122223333
    UserInfo:
      type: array
      items:
        type: object
        properties:
          user_id:
            type: integer
            example: 1
          username:
            type: string
            example: ghee_dough
          full_name:
            type: string
            example: Guido Torres
          email:
            type: string
            example: gheedough@gmail.com
          phone:
            type: integer
            example: 5491122223333
          delivery_address:
            type: string
            example: Fake St. 123
          is_admin:
            type: boolean
            example: 0
          is_disabled:
            type: boolean
            example: 0
...
    Token:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
    Error:
      type: object
      properties:
        error:
          type: string
          example: "An error has ocurred: <error>"
      required:
        - error
    Send:
      type: string
      example: "Error message"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Reused Responses
  responses:
...
    Disabled:
      description: Invalid request, user account is disabled
...
    NotFound:
      description: The specified resource was not found
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Send"
          examples:
            Token Invalidation:
              summary: Unauthorized - Invalid Token
              value: Unauthorized - Invalid Token
            Disabled Account:
              summary: Unauthorized - Account is Disabled
              value: Unauthorized - User account is disabled
...
security:
  - bearerAuth: []